+++ opentelemetry-sdk/src/opentelemetry/sdk/_shared_internal/_processor_metrics.py
@@ -19,6 +19,12 @@ from collections.abc import Callable
 from typing import Literal
 
 from opentelemetry.metrics import CallbackOptions, MeterProvider, Observation
+from opentelemetry.sdk.environment_variables import (
+    OTEL_PYTHON_SDK_METRICS_ENABLED,
+)
+from opentelemetry.sdk.environment_variables._internal import (
+    parse_boolean_environment_variable,
+)
 from opentelemetry.semconv._incubating.attributes.otel_attributes import (
     OTEL_COMPONENT_NAME,
     OTEL_COMPONENT_TYPE,
@@ -75,12 +81,19 @@ class ProcessorMetrics:
             )
 
         self._processed = create_processed(meter)
+        self._enabled = parse_boolean_environment_variable(
+            OTEL_PYTHON_SDK_METRICS_ENABLED
+        )
 
         if capacity is not None:
             self._queue_capacity = create_queue_capacity(meter)
-            self._queue_capacity.add(capacity, self._standard_attrs)
+            if self._enabled:
+                self._queue_capacity.add(capacity, self._standard_attrs)
 
     def register_queue_size(self, get_queue_size: Callable[[], int]) -> None:
+        if not self._enabled:
+            return
+
         def record_queue_size(
             _options: CallbackOptions,
         ) -> tuple[Observation]:
@@ -103,9 +116,13 @@ class ProcessorMetrics:
         )
 
     def drop_items(self, count: int) -> None:
+        if not self._enabled:
+            return
         self._processed.add(count, self._dropped_attrs)
 
     def finish_items(self, count: int, error: Exception | None) -> None:
+        if not self._enabled:
+            return
         if not error:
             self._processed.add(count, self._standard_attrs)
             return