+++ opentelemetry-sdk/tests/metrics/test_periodic_exporting_metric_reader.py
@@ -20,10 +20,13 @@ import weakref
 from logging import WARNING
 from time import sleep, time_ns
 from typing import Optional, cast
-from unittest.mock import Mock
+from unittest.mock import Mock, patch
 
 import pytest
 
+from opentelemetry.sdk.environment_variables import (
+    OTEL_PYTHON_SDK_METRICS_ENABLED,
+)
 from opentelemetry.sdk.metrics import (
     Counter,
     MeterProvider,
@@ -315,6 +318,7 @@ class TestPeriodicExportingMetricReader(ConcurrencyTestBase):
             "The PeriodicExportingMetricReader object created by this test wasn't garbage collected",
         )
 
+    @patch.dict("os.environ", {OTEL_PYTHON_SDK_METRICS_ENABLED: "true"})
     def test_metric_reader_metrics(self):
         exporter = FakeMetricsExporter()
         pmr = PeriodicExportingMetricReader(