+++ opentelemetry-sdk/tests/trace/test_sdk_metrics.py
@@ -13,8 +13,12 @@
 # limitations under the License.
 
 from unittest import TestCase
+from unittest.mock import patch
 
 from opentelemetry import trace as trace_api
+from opentelemetry.sdk.environment_variables import (
+    OTEL_PYTHON_SDK_METRICS_ENABLED,
+)
 from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.metrics.export import InMemoryMetricReader
 from opentelemetry.sdk.trace import TracerProvider
@@ -27,6 +31,7 @@ from opentelemetry.sdk.trace.sampling import (
 from opentelemetry.trace.span import SpanContext
 
 
+@patch.dict("os.environ", {OTEL_PYTHON_SDK_METRICS_ENABLED: "true"})
 class TestTracerProviderMetrics(TestCase):
     def setUp(self):
         self.metric_reader = InMemoryMetricReader()
@@ -159,6 +164,7 @@ class TestTracerProviderMetrics(TestCase):
             },
         )
         self.assert_live_spans(metric_data, None, {})
+
         span.end()
         metric_data = self.metric_reader.get_metrics_data()
         self.assert_started_spans(
@@ -242,3 +248,19 @@ class TestTracerProviderMetrics(TestCase):
             },
         )
         self.assert_live_spans(metric_data, None, {})
+
+
+class TestTracerProviderMetricsDisabled(TestCase):
+    def test_disabled_by_default(self):
+        metric_reader = InMemoryMetricReader()
+        meter_provider = MeterProvider(metric_readers=[metric_reader])
+        tracer_provider = TracerProvider(
+            sampler=ALWAYS_ON, meter_provider=meter_provider
+        )
+        tracer = tracer_provider.get_tracer("test")
+
+        with tracer.start_as_current_span("span"):
+            pass
+
+        self.assertIsNone(metric_reader.get_metrics_data())
+        meter_provider.shutdown()